+++ packages/predictive-search/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bryt-designs/predictive-search
 
+## 3.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @bryt-designs/storefront-client@0.7.5
+
 ## 3.0.4
 
 ### Patch Changes